Add files via upload
After I finished my previous project, I was eager to take and learn something new. Since the previous project inevitably took so much time to complete, I was looking for something simpler. I wanted to figure out a way to develop an app for the weather forecast for a selected city. OpenWeatherMap was the source for my API, which deserialized into readable JSON files every 15 seconds and displayed them on the main form. It provided me with more data than I knew what to do with, and lots of potential functionalities to implement. The final result is an app that allows you time, date and current weather data in the main interface with ability to search for city you want to view data for.
•	API’s,
•	asynchronous tasks,
•	using settings file to save last succesfully fetched city

Maintainability Index: 81 Cyclomatic Complexity: 34 Depth of Inheritance: 7 Class Coupling: 38 Lines of Source code: 468 Lines of Executable code: 203